Restructured s840dplugin for better performance.
SublimeText's API is very slow, which causes the operations to take several seconds on larger files.
Therefore the whole content is now read from sublime's buffer and operations are performed
on the text directly. Finally the whole text is pushed back to the view buffer.

Renumbering is now done without regexp to maximize performance. Altogether this speeds up
renumbering by a factor of 1000.

The only remaining bottleneck is the update of sublime's text buffer, which takes several seconds
for larger files with around 10.000 or more lines.