Other: Update syntax specific files with <template>.sublime-settings
Instead of hard coding essential settings, they are again stored in usual
sublime-settings files. As ST's load_settings() sometimes ignores settings
files whose name matches a source scope, they need to be loaded and copied
to the syntax settings file manually. This also ensures essential settings
to be applied for the three syntaxes.

The code is moved to the plugin.py.