RunMyScreens: Enable auto-completion of sofktey and column attributes.
The syntax defines scope based auto-triggers to show auto-completion
of attributes. If a scope is popped off, the matching pattern '\)' does not
include the meta_content_scope anymore. This prevents autocompetion to work
properly, in some edge cases. If the next scope is 'set', the '\)' contains
the meta_content_scope and auto-completion works.