G-Code: Optimize {{cycle}} pattern for better performance
Remove capture and look-ahead from {{cycle}} as it is not needed in most usages.
Instead query the look-ahead only in direct cycle calls, which really need it.
This improves overall scoping performance in files with heavy cycle usage.

In general it is a good practice to avoid captures at all as they are expensive.