G-Code: Ensure bailout in multi-cycle archives
If an archive file contains several procedures/cycles an syntax error may have caused the `%` archive header of the next cycle to have been ignored up to the previous commit.

Therefore the `^%` pattern is used as prototype to ensure it is always matched first with highest possible priority, ignoring the context of the included syntax file.