G-Code: Further fixes for some edge cases of the minify command
The G-Code syntax does currently not include newlines in comments if they are part of a macro definition due to the way we need to escape from macros. Hence the minify command breaks the code then.

The `meta.eol` is now added to each captured newline and used to capture comments without it.